Don't consume objects in public API
Since javascript doesn't have a borrow checker, functions that consume
their input are easy to misuse and don't behave in a way that's easy
to understand in JS. Calling a seemingly pure function can nullify one
of its arguments.

This introduces a few more clones, but I think the improved API makes
it worth it.